Moved tests, fixed imports, and renamed CoverageDropoutDetectorTest.
